version 2.0.0-SNAPSHOT-1 Added integration for Gradle and GitHub Packages, has to be tested. Re-framed the tree of src -> moved to library/src/io.github.justlel.tgbot Added sample with configs, README must be updated.
