+++ gradle.properties
@@ -1,0 +1,11 @@
+GROUP=io.github.justlel
+VERSION_NAME=2.0.0-SNAPSHOT-1
+POM_DESCRIPTION=An integration to pengrad's telegram-bot-api library that helps to handle updates received by the bot.
+POM_URL=https://github.com/justlel/tgbot-handlers-integration/
+POM_SCM_URL=https://github.com/justlel/tgbot-handlers-integration/
+POM_SCM_CONNECTION=scm:git:git://github.com/justlel/tgbot-handlers-integration.git
+POM_SCM_DEV_CONNECTION=scm:git:git://github.com/justlel/tgbot-handlers-integration.git
+POM_LICENCE_NAME=GNU General Public License v3.0
+POM_LICENCE_URL=https://www.gnu.org/licenses/gpl-3.0.txt
+POM_LICENCE_DIST=repo
+POM_DEVELOPER_ID=justlel
