+++ library/build.gradle
@@ -1,0 +1,20 @@
+plugins {
+    id 'java'
+    id 'com.github.johnrengelman.shadow' version '7.1.2'
+}
+
+version '2.0.0-SNAPSHOT-1'
+
+
+dependencies {
+    shadow localGroovy()
+    shadow gradleApi()
+
+    implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.13.2'
+    implementation group: 'com.github.pengrad', name: 'java-telegram-bot-api', version: '6.0.1'
+    testImplementation group: 'junit', name: 'junit', version: '4.13'
+}
+
+test {
+    useJUnit()
+}
