+++ samplebot/src/main/resources/sample-configs.yml
@@ -1,0 +1,1 @@
+token: ""
