+++ settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = "tgbot-handlers-integrtion"
+
+include 'library', 'samplebot'
